+++ lib/interactive_3d_method_channel.dart
@@ -1,11 +1,9 @@
 import 'dart:async';
-
-import 'package:flutter/foundation.dart';
+import 'dart:typed_data';
 import 'package:flutter/services.dart';
-
+import 'package:flutter/foundation.dart';
 import 'interactive_3d_platform_interface.dart';
 
-/// An implementation of [Interactive3dPlatform] that uses method channels.
 class MethodChannelInteractive3d extends Interactive3dPlatform {
   final MethodChannel _methodChannel;
   final EventChannel _eventChannel;
@@ -16,17 +14,18 @@ class MethodChannelInteractive3d extends Interactive3dPlatform {
     _eventChannel.receiveBroadcastStream().listen(_onEvent);
   }
 
-  final StreamController<List<EntityData>> _selectionController = StreamController.broadcast();
+  final StreamController<List<EntityData>> _selectionController =
+  StreamController.broadcast();
 
   @override
   Stream<List<EntityData>> get selectionStream => _selectionController.stream;
 
   void _onEvent(dynamic event) {
-    final Map<dynamic, dynamic> map = event;
+    final map = event as Map<dynamic, dynamic>;
     final String eventType = map['event'];
     if (eventType == 'selectionChanged') {
       final List<dynamic> selectedEntities = map['selectedEntities'];
-      final List<EntityData> entities = selectedEntities.map((e) {
+      final entities = selectedEntities.map((e) {
         return EntityData(id: e['id'], name: e['name']);
       }).toList();
       _selectionController.add(entities);
@@ -34,31 +33,42 @@ class MethodChannelInteractive3d extends Interactive3dPlatform {
   }
 
   @override
-  Future<void> loadModel(String modelPath) async {
-    ByteData data = await rootBundle.load(modelPath);
-    Uint8List modelBytes = data.buffer.asUint8List();
-    await _methodChannel.invokeMethod('loadModel', {'modelBytes': modelBytes});
+  Future<void> loadModel(String modelPath, Map<String, ByteData> resources) async {
+    // Convert the main model to bytes.
+    ByteData modelData = await rootBundle.load(modelPath);
+    Uint8List modelBytes = modelData.buffer.asUint8List();
+
+    // Convert the resources map to a sendable form.
+    final resourceMap = resources.map(
+          (key, value) => MapEntry(key, value.buffer.asUint8List()),
+    );
+
+    final args = {
+      'modelBytes': modelBytes,
+      'name': modelPath.split('/').last,
+      'resources': resourceMap
+    };
+
+    await _methodChannel.invokeMethod('loadModel', args);
   }
 
   @override
-  Future<void> loadEnvironment(String? iblPath, String? skyboxPath) async {
+  Future<void> loadEnvironment(String iblPath, String skyboxPath) async {
     Uint8List? iblBytes;
     Uint8List? skyboxBytes;
 
-    if (iblPath != null && skyboxPath != null) {
-      try {
-        ByteData iblData = await rootBundle.load(iblPath);
-        iblBytes = iblData.buffer.asUint8List();
-      } catch (e) {
-        print('Error loading IBL file: $e');
-      }
+    try {
+      ByteData iblData = await rootBundle.load(iblPath);
+      iblBytes = iblData.buffer.asUint8List();
+    } catch (e) {
+      debugPrint('Error loading IBL file: $e');
+    }
 
-      try {
-        ByteData skyboxData = await rootBundle.load(skyboxPath);
-        skyboxBytes = skyboxData.buffer.asUint8List();
-      } catch (e) {
-        print('Error loading skybox file: $e');
-      }
+    try {
+      ByteData skyboxData = await rootBundle.load(skyboxPath);
+      skyboxBytes = skyboxData.buffer.asUint8List();
+    } catch (e) {
+      debugPrint('Error loading skybox file: $e');
     }
 
     await _methodChannel.invokeMethod('loadEnvironment', {