+++ lib/interactive_3d_platform_interface.dart
@@ -1,26 +1,25 @@
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 import 'dart:async';
+import 'dart:typed_data';
 
 abstract class Interactive3dPlatform extends PlatformInterface {
-  /// Constructs a Interactive3dPlatform.
   Interactive3dPlatform() : super(token: _token);
 
   static final Object _token = Object();
 
-  /// Verify the instance
   static void verify(Interactive3dPlatform instance) {
     PlatformInterface.verifyToken(instance, _token);
   }
 
-  // Method to load the model
-  Future<void> loadModel(String modelPath);
+  /// Method to load the model.
+  /// [resources] is empty for .glb models and populated for .gltf models.
+  Future<void> loadModel(String modelPath, Map<String, ByteData> resources);
 
-  // Method to load the environment
+  /// Method to load the environment.
   Future<void> loadEnvironment(String iblPath, String skyboxPath);
 
-  // Stream to receive selection changes
+  /// Stream to receive selection changes.
   Stream<List<EntityData>> get selectionStream;
-
 }
 
 class EntityData {