+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Initial release of `interactive_3d`.
+- Features:
+    - Render `.glb` and `.gltf` models.
+    - Interactive touch gestures to extract information.
+    - Currently supports only Android using the Filament Engine.
+    - Open-source and customizable.